Revise and extend drag/drop functionality
 - Add TopmostParent method to GuiWidget
 - Reuse TopmostParent in TransformFromScreenSpace and drag/drop
 - Consolidate file/mesh loading implementations
 - Bust out Object3D classes into separate files
 - Move all Object3D types to a common assembly
