+++ DataConverters3D/DataConverters3D.csproj
@@ -139,8 +139,14 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CsgToRayTraceable.cs" />
+    <Compile Include="Object3D\InteractiveScene.cs" />
+    <Compile Include="Object3D\Json\MatrixConverter.cs" />
+    <Compile Include="Object3D\Json\Object3DConverter.cs" />
+    <Compile Include="Object3D\PlatingData.cs" />
+    <Compile Include="Object3D\Object3DExtensions.cs" />
+    <Compile Include="Object3D\IObject3D.cs" />
     <Compile Include="MeshToBVH.cs" />
-    <Compile Include="Object3D.cs" />
+    <Compile Include="Object3D\Object3D.cs" />
     <Compile Include="VertexSourceToMesh.cs" />
   </ItemGroup>
   <ItemGroup>
