+++ examples/MeshViewer/MeshViewer.csproj
@@ -251,4 +251,5 @@
     <None Include="app.config" />
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
 </Project>
